requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.15.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
bcrypt==4.3.0
    # via passlib
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via
    #   typer
    #   uvicorn
cryptography==44.0.2
    # via python-jose
dynaconf==3.2.10
    # via -r requirements.in
ecdsa==0.19.1
    # via python-jose
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.6
    # via -r requirements.in
mako==1.3.9
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.1
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.33.0
    # via pydantic
pygments==2.19.1
    # via rich
python-jose[cryptography]==3.4.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
rich==14.0.0
    # via
    #   -r requirements.in
    #   typer
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.24
    # via -r requirements.in
starlette==0.46.1
    # via fastapi
typer==0.15.2
    # via -r requirements.in
typing-extensions==4.13.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich
    #   sqlalchemy
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.0
    # via -r requirements.in
